Atualizado pagina inicial

+++ lib/Models/produto.dart
@@ -3,7 +3,7 @@ class Produto {
   int? _codigo_categoria;
   String? _nome;
   String? _descricao;
-  int? _preco;
+  double? _preco;
 
   //construtor
   Produto({
@@ -11,7 +11,7 @@ class Produto {
     int? codigo_categoria,
     String? nome,
     String? descricao,
-    int? preco,
+    double? preco,
   }) {
     _codigo_do_produto = codigo_do_produto;
     _codigo_categoria = codigo_categoria;
@@ -23,7 +23,7 @@ class Produto {
   //getters tradicional
   int? get codigo_do_produto => _codigo_do_produto;
   int? get codigo_categoria => _codigo_categoria;
-  int? get preco => _preco;
+  double? get preco => _preco;
   String? get nome => _nome;
   String? get descricao => _descricao;
 
@@ -32,7 +32,7 @@ class Produto {
       _codigo_do_produto = codigoDoProduto;
   set codigo_categoria(int? codigoCategoria) =>
       _codigo_categoria = codigoCategoria;
-  set preco(int? preco) => _preco = preco;
+  set preco(double? preco) => _preco = preco;
   set nome(String? nome) => _nome = nome;
   set descricao(String? descricao) => _descricao = descricao;
 }
